backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.10.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
argon2-cffi==23.1.0
slowapi==0.1.9
# Removed pyjwt - using session-based auth instead of JWT
